data_generator/requirements.txt
loguru==0.7.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
Faker==19.3.1